requirement.txt
django
djangorestframework
django-qr-code
gnosis-py
coverage
web3
py-solc-x
